test: Move iUserSpace tests to iUserSpaceFunctional.js
